functions/makekml/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery
google-cloud-storage
simplekml

